move multimodal and refactor block setting
