+++ vllm_rbln/utils/optimum/compilation/multimodal/idefics3.py
